feat(material-requests): generate graphql types
