+++ src/modules/material-requests/material-requests.consts.ts
@@ -1,4 +1,4 @@
-import { MaterialRequest } from '~modules/material-requests/types';
+import { MaterialRequest } from '~modules/material-requests/material-requests.types';
 
 export const ORDER_PROP_TO_DB_PROP: Partial<Record<keyof MaterialRequest, string>> = {
 	id: 'id',
