+++ src/modules/material-requests/services/material-requests.service.ts
@@ -3,9 +3,14 @@ import { Injectable, Logger } from '@nestjs/common';
 import { IPagination, Pagination } from '@studiohyperdrive/pagination';
 import { isEmpty, set } from 'lodash';
 
-import { ORDER_PROP_TO_DB_PROP } from '../consts';
+import {
+	FindMaterialRequestsDocument,
+	FindMaterialRequestsQuery,
+	FindMaterialRequestsQueryVariables,
+} from '../../../generated/graphql-db-types-hetarchief';
 import { MaterialRequestsQueryDto } from '../dto/material-requests.dto';
-import { MaterialRequest } from '../types';
+import { ORDER_PROP_TO_DB_PROP } from '../material-requests.consts';
+import { GqlMaterialRequest, MaterialRequest } from '../material-requests.types';
 
 import { PaginationHelper } from '~shared/helpers/pagination';
 import { SortDirection } from '~shared/types';
@@ -22,11 +27,11 @@ export class MaterialRequestsService {
 	public adapt(grapqhQLMaterialRequest: GqlMaterialRequest): MaterialRequest | null {
 		return {
 			id: grapqhQLMaterialRequest.id,
-			objectSchemaIdentifier: grapqhQLMaterialRequest.objectSchemaIdentifier,
-			profileId: grapqhQLMaterialRequest.profileId,
+			objectSchemaIdentifier: grapqhQLMaterialRequest.object_schema_identifier,
+			profileId: grapqhQLMaterialRequest.profile_id,
 			reason: grapqhQLMaterialRequest.reason,
-			createdAt: grapqhQLMaterialRequest.createdAt,
-			updatedAt: grapqhQLMaterialRequest.updatedAt,
+			createdAt: grapqhQLMaterialRequest.created_at,
+			updatedAt: grapqhQLMaterialRequest.updated_at,
 		};
 	}
 
@@ -43,12 +48,12 @@ export class MaterialRequestsService {
 		const where: FindMaterialRequestsQueryVariables['where'] = {};
 
 		if (!isEmpty(parameters.userProfileId)) {
-			where.user_profile_id = { _eq: parameters.userProfileId };
+			where.profile_id = { _eq: parameters.userProfileId };
 		}
 
 		const materialRequestsResponse = await this.dataService.execute<
 			FindMaterialRequestsQuery,
-			FindMaterialRequestsVariables
+			FindMaterialRequestsQueryVariables
 		>(FindMaterialRequestsDocument, {
 			where,
 			offset,
