--- src/modules/material-requests/types.ts
@@ -1,8 +1,0 @@
-export interface MaterialRequest {
-	id: string;
-	objectSchemaIdentifier: string;
-	profileId: string;
-	reason: string;
-	createdAt: string;
-	updatedAt: string;
-}
